Updated openstack/openstack Project: openstack/horizon  fed697cab5698f4f5edaa941117976ca6b3bbe67
passing wrong dict in lbaas

While updating the member and health monitor in lbass,
the resultant dict passing to the constructor is not uniform with the other methods(get and create).

This will cause a KeyError, if we access the dict values.

This patch making the dict uniform with all the methods.

Change-Id: I8b58a0acb4153358b9ed4a46794aaccce181e31b
Closes-Bug: #1438533