Updated openstack/openstack Project: openstack/glance  80a8ff330fcb633a8aa2d577dd47a274a9fd9a58
A mixin for jsonpatch requests validation

A JSONPatch-enabled API requires validation logic which will check the
correctness of incoming request body according to schema. A mixin is
added which will do such validations.

Such validation logic may be useful not only for Artifacts API but also
for regular v2 images API, as it also uses JSONPatch notation in its
update calls. So this mixin may be used in future in Images API as well.

Implements-blueprint: artifact-repository
Change-Id: I1b7b72bcc0a0455803a4e99dd17a9732b0dd7a8d