Updated openstack/openstack Project: openstack/openstackdocstheme  1eabac2ed20edc399fed56a00ffaa516ae4d279c
Add the output from pygmentize

Currently, the tango pygments_style theme is set to tango,
but there was no output from

pygmentize -S tango -f html

Which is required to create the css file.

Tango is does not render nicely, so this patch also switches to the
native theme, which is readable.

Change-Id: Ie69486ba7269d2f4b4adb6d17bb205c2f7cd0161
Closes-bug: #1439523