Updated openstack/openstack Project: openstack/horizon  73275906ae17cf39d7fabfdce1b6e0bca9d5c0d2
Fix sort order for User IDs in identity table

The order that the user IDs are sorted into by the default tablesorter
parsers is incorrect, and seems to be getting confused by the
complexity of the values in these fields. This patch adds a new
custom parser that performs some logic to normalize these values
correctly and in a way that returns the correct sorted order.

Change-Id: I62f91ff4ca4e9110d7db5b66cdbf4b3ea15105c6
Closes-Bug: 1438818