Updated openstack/openstack Project: openstack-dev/devstack  5cb3d6b7eeed5b1fb6cdd736a7186240dafdddf2
Install packaged pyOpenSSL

pyOpenSSL has done a rewrite of itself in Python.  This may be good
for many reasons, but memory usage is not one of them.  It just about
doubles the size of at least swift, which usually consumes about 6% of
a CI testing vm's 8gb RAM.  This is enough to push centos hosts into
OOM conditions and then everything falls apart.

The distro packages of pyOpenSSL are the older C-based versions, which
doesn't bring in the kitchen sink of cffi & pycparser.

Change-Id: Icd4100da1d5cbdb82017da046b00b9397813c2f2