Updated openstack/openstack Project: openstack-infra/shade  7a68f2710939e459250d9239b16f47fc4399952e
Add task management framework

Some shade clients, like nodepool, have a need to manage threads that
are doing API tasks and to manage rate limiting. That's not really a
thing that's appropriate for shade to know about, but because shade does
want to know the business logic, we need to provide a mechanism for
people to do that.

TaskManager is essentially nodepool.TaskManager except with
threading features removed and reworked to do instantaneous blocking
execution, since that's the behavior that most users will expect.

A patch will follow to move API calls to be manged by Tasks and the
shade.TaskManager. Once those are there, then nodepool can pass in its
TaskManager and the API operations in shade will naturally be managed by
the rate-limited operations in nodepool.

Co-Authored-By: James E. Blair <jeblair@openstack.org>
Change-Id: I60d25271de4009ee3f7f7684c72299fbd5d0f54f