Updated openstack/openstack Project: openstack-infra/shade  8be85afd2871e98e401112fdf69181e3afd7e516
Create swift container if it does not exist

In the glance v2 image upload workflow, the user is not doing things
with swift containers, swift containers are being created by shade. This
means that it's shade's responsibility to ensure that the container
exists before it attempts to upload something to it.

Change-Id: I34cfc2821819aec105947dbe70a578307e583d77