Updated openstack/openstack Project: openstack-infra/project-config  282cdf0dd082428002f4ef00085ac6ad83171f4b
Upload devstack-gate documentation with logs

We need to place the tempest log documentation in with the devstack
logs. We can do this as a publisher shell step.

The tempest overview is a little more complicated as it needs
placing in the top most job results directory. We can scp this for
the old log storage and then for zuul-swift-logs we pass it as a
file to be uploaded which will hence be in the master index (ie at
the top of our pseudo filesystem)

Change-Id: I54d2d2ba9b6532221ca3b376ab2e51bc5398444e
Depends-On: I8b26cc58c2ad6ca029726d227cf1d38d72465627