Updated openstack/openstack Project: openstack/nova  6871585b2462cbce6096d4366c1b956cca6545ee
Be more forgiving to empty context in notification

We seem to call
rpc.get_notifier('api').error(None, 'api.fault', payload) in
notifications.py, where the first parameter is context. This
eventually calls context.to_dict() in nova/rpc.py, which breaks
for a NoneType object. So we should pick the current context
from oslo_context or switch to the admin context.

Co-Authored-By: Pranesh Pandurangan <praneshpg@gmail.com>
Closes-Bug: 1360911
Change-Id: Ic37e184a56377739cd7cda5ba4bde236990fef9e