Updated openstack/openstack Project: openstack/neutron  a38d294f1e1ad515d86952e297d2b4d437d99cff
Remove auto deletion of routers in unit tests

Remove the automatic deletion behavior of the router
context manager in the L3 unit tests. Any tests that
depend on the router being deleted should do so
explicitly.

It additionally removes the logic from the test_l3_plugin
unit tests that was just related to tearing down enough
stuff to allow the context managers to exit. It was code
that distracted from what the tests were actaully verifying.

All of the context managers for port, network, and subnet
do not auto delete by default and that will be extended to
the L3 constructs as well. The patch that did this for
ports/subnets/networks is here:
https://review.openstack.org/#/c/102465/

Change-Id: Iec97198f18e9fc390ff0747b795f7f309c8f3990