Updated openstack/openstack Project: openstack-dev/devstack  53ded8cedaca91c1ad576b81b090b7bc237b4a8f
minimize the default services

This changes the default service list in devstack to minimize what is
running out of the box, so that it's likelihood of working in a 4G vm
is much higher.

This removes heat from the default enabled service list.

It drops the ec2 only needed n-obj and n-crt services.

It drops all the alternative consoles (xvnc, consoleauth). novnc is
fine for libvirt which is the default.

It adds dstat, because that's turned out to be so useful in debugging
things.

Change-Id: I84457260dff6f42a5c6ebcc2c60fb6e01aec9567