Updated openstack/openstack Project: openstack/manila  a327dcfb500176fffbdef099e89d8daa354c4a47
cDOT driver should split clone from snapshot after creation

When the cDOT driver receives a create_share_from_snapshot command,
it clones the share/flexvol from the specified snapshot in the
parent share/flexvol.  But if the resulting clone is not split
from its parent, the snapshot remains busy and may not be deleted.
This is resolved by starting a clone split operation anytime a
clone is created.

During the split operation, which usually completes within
seconds, the snapshot remains busy.  This patch detects that
condition in the snapshot delete path and waits up to a minute
for the busy condition to clear.

Change-Id: I44bac106f7f7a1acc968c1d9a526dcef85f657b7
Closes-Bug: #1259988