Updated openstack/openstack Project: openstack/openstack-manuals  48fa375ea8b61e3d4387315af9839b2122edf535
Domains with users and projects

Administrators need to know about domains along with users and
projects, so move the description of domains into the "User
management" section where users and projects are defined.
Considering how short the domains section was it fits in.

Not mentioning domains along with users and projects leads to
confusion.

Change-Id: I8117677a3bbf96f1caa6f17017e077391ec64457