Updated openstack/openstack Project: openstack/ironic  54710214a41f1a98000431e8a5f2685af959cc58
Fix is_hostname_safe for RFC compliance

In the move from Node name supporting just a hostname to now
supporting a FQDN, we missed a few changes to keep this function
being RFC compliant.  This patch enhances the code and tests for
this, and migrates the database to the longer name field format.

Change-Id: Ib75c8b138d7165aedc74efead80d3fb755cab87b
Closes-Bug: #1433832