Updated openstack/openstack Project: openstack/nova  277ea04467c2fb0f24716b8e7e7799b43fc50af0
Fix incorrect statement in inline neutronv2 docs

The function doc header for network/neutron/v2/api.py _create_port()
states that it accepts an optional set of MAC addresses to use.

However, it only uses one, which is selected by pop().

This commit merely changes the doc string to indicate that the function
operates on only a single MAC address, not on the whole set.

Change-Id: I0fc845572836b6a6773cc7beaeefd7b0d1cfd5e7
Related-bug: #1405131