Updated openstack/openstack Project: openstack/keystone  3d6631d08de5e248472e55d42030c1d25e271503
Cleanup use of .driver

Unless being used to override the Manager implementation, .driver
doesn't need to be used because the Manager automatically
forwards calls to the driver. The code shouldn't use .driver
directly where it's not necessary since developers should be
able to override behavior in the Manager, which we've done
several times (for deprecations or reorganization for
example).

Change-Id: I05b13d4f4ec7e6be8e72b73bac164333535aebee