Updated openstack/openstack Project: openstack/oslo.serialization  6216d7b61045800067aed7434c65f59b4235d970
Make the msgpackutils handlers more extendable

This change revamps the current serialization and
deserialization functions to be tiny objects that
have the ability to serialize there expected types.

It also adds a registry of these handlers objects
and allows for passing a registry into the dumps
and loads methods for people that want to add there
own handlers on.

Change-Id: Ib1a8a507c6c3d4f0372114d67163d3f2ab495c9f