Updated openstack/openstack Project: openstack/oslo.policy  a62b7e00d7a02d10d7f9c7acfb649b616564760a
Lists for Generic Checks

The Generic check had no way to identify
a value inside a list.  Since Lists are inherantly unstable in
indexing, requiring an index to match is not practical.

Lists now follow the same approach as the OrCheck:
When specifying a value inside a list, each element
of the lsit is checked for a match.  If any of the sub lists match,
the check succeeds. Only if the entry is not in the list does the
check fail.

If the value is nested in a dictionary under the list, all of
the subordinate dictionaries are checked in a recursive manner.

Change-Id: Ia286dbd3757703779d7044b3003381eab6c5c919