Updated openstack/openstack Project: openstack-dev/devstack  70843539a4ab2ed3c4b5d7042cc5ea717407e488
Set qcow image file location with right cirros version

This commit ensures we also set the qcow image location in the
tempest config when we update qcow version. The tempest config has
a default value for img_file (which is incorrect) but before we can
remove the defaults in tempest we need to ensure devstack is using
it properly first. The only reason the tests weren't failing here is
because tempest falls back to using uec images (which devstack was
correctly setting config for) if qcow isn't found. The img_dir was
also hardcoded assuming a uec image, however if qcow is intended to be
used you'll need to be able to override that, which is added as
part of this commit.

Change-Id: I05af346b3c9be9560dc8846dd1f437cfbb2d5005