Updated openstack/openstack Project: openstack-dev/devstack  12909c0a76d29261ebd20357ffa5a70ed65afd91
If cells enabled create a fixed network in the API cell

Now that tempest is querying for a network before making server create
calls the fixed network needs to be known by the API cell.  Server
creates should work for networks defined in both databases, but defining
a new network via the API will continue to not work.

Change-Id: I32461add0d20940a55385c8b34cd493e2561615e