Updated openstack/openstack Project: openstack/ironic  8ee60f629c4719b947d57b2a76afe26d60ec19b0
Improve error handling when JSON is not returned by agent

Sometimes the agent is not returning valid JSON in the response.  Handle
it better and record information about the request and the response to
help with debugging.

Will now raise an exception similar to this:
ironic.common.exception.IronicException: Unable to decode response as JSON.
Request URL: http://127.0.0.1:9999/v1/commands
Request body: "{"name": "standby.run_image",
                "params": {"image_info": {"image_id": "test_image"}}}"
Response: "this is not json"

Change-Id: I483884ca5727fbaa3120e2f7c2b381a7d7e8be17
Closes-Bug: 1439498