Updated openstack/openstack Project: openstack/cinder  7ba197ae285b20b6a62365f3374f265a2dbbeff6
Only use operational LIFs for iscsi target details

When Nova invokes our cDOT driver's initialize_connection() method
in order to attach an iSCSI volume to an instance, we have been
returning the first target from the list of targets returned from
the filer API call to get iscsi target details.  In failover mode,
this may be a target with a LIF that is down, causing the attach
attempt to fail in Nova.

This commit fixes that issue so that attaches still work in
failover by returning the first target with an *operational* LIF.

Closes-bug: 1437419

Change-Id: I3f53c3644f7d94f5188151c16d70e565b745ad4a