Updated openstack/openstack Project: openstack/manila  52750a0824d3724f9e2887a7090f9c4ab6111b00
cDOT multi-SVM driver may choose unsuitable physical port for LIFs

The algorithm used by the NetApp cDOT multi-SVM driver for choosing
a physical port on which to create logical interfaces (LIFs) is too
simplistic, which can result in the selection of ports that won't
work at all or which are far inferior to other choices.  This patch
improves the port selection query, sorts ports by speed, and allows
the user to override the port selection via a regex setting in the
manila.conf file.

Closes-Bug: #1434918
Change-Id: Idedcb0db23df3a745f0c1ced2652da008746f0a6