Updated openstack/openstack Project: openstack/keystone  384c3f9100c4a69fe4300a1db9cd5cb233d1ed60
Allow identity provider to be created with remote_ids set to None

When support was added for multiple remote_ids, identity provider
creation was broken when remote_ids are specified as None instead
of an empty list.  The remote_ids are supposed to be optional.  This
causes 400 errors when using python-openstackclient to create an
identity provider without any remote_ids.

Change-Id: I9a4d1206f6df95636ae350236f6e704ca37f7b73
Closes-bug: #1440185