Updated openstack/openstack Project: openstack/nova  79fe073c9938dbd32550407c4201ea0bb8366ad0
virt: Make sure block device info is persisted

There are several methods that we wrap with a @update_db decorator,
which is to say that we want the data that changed on the object to be
persisted once the method has done it's work.

However in case of exceptions, we would do some local cleanup and
propagate the exception to through the stack, which would cause the db
save to not happen (due to the way the 'update_db' decorator was
written).

This is not ideal since some of those methods such as 'attach()' do
several interaction (with the Cinder API and libvirt for example), and
making sure that results of successfull ones are persisted in case of
a non immediate failure is important for later cleanup, and keeping
the system in a consistent state.

This patch makes sure that we persist data to the DB at all times.

Change-Id: Icf8dbe4e3b7fd65b3c7a5eb479d6be15ee504253
Closes-bug: #1439282