Updated openstack/openstack Project: openstack/neutron  bcaca1c3c23c94dfaf0ac64c8bc6cd85a6cbe06e
Make L3 agent honor periodic_interval setting

The periodic_task decorator for the sync routers task was resulting
in a default spacing of 60 seconds. This meant that any values less
than that for the periodic_interval setting would not work correctly.
The fixed interval looping call would run at the periodic_interval
but this task would not execute every time as expected.

For example, if the periodic_interval was 40 seconds, the task would
only end up running every other interval (80 seconds in this case)
because every other attempt would be blocked by the default 60 second
barrier of periodic_task.

This sets the periodic_taks spacing variable to 1 second so the
interval is controlled only by the loopingcall as expected.

Ultimately periodic_task should probably be completely removed since
it's not compatible with the fixed interval loopingcall in this manner.

Closes-Bug: #1411085
Change-Id: I23818c3fab2640b241692f00f9b5a2f923e3cf31