Updated openstack/openstack Project: openstack/nova  a4c811f5403fbd44c93e98a6a41c792c0abf84a8
Respect CONF.scheduler_use_baremetal_filters

This fixes an issue where the default filters are loaded before
IronicHostManager overrides the default, which is the intended behavior
when CONF.scheduler_use_baremetal_filters is True.

Creates a _load_filters method on the base HostManager class that
returns a list of filters to be loaded at initialization, and overrides
it in IronicHostManager to return the default baremetal filters when
scheduler_use_baremetal_filters is set.

Change-Id: I1801c62d85665a6a3423428e91e1aeeb546613d7
Closes-Bug: #1439868