Updated openstack/openstack Project: openstack/cinder  df7ed172e12c782a85394a1526adb0aec0b6aa37
Fix ISCSIDriver initialized connection volume type

ISCSIDriver initialize connection is a duplicated function
left in kilo to serve drivers that don't use the new targets model.

In Kilo there is a new parameter of iscsi_protocol, that is set to
iscsi by default and can be changed to iser in order to enable RDMA.

In order support RDMA in drivers that still inherit from ISCSIDriver,
driver_volume_type should be set to iscsi_protocol parameter value
instead of a hard coded value until this function duplication will be solved.

Closes-Bug: #1438833

Change-Id: If51ff933b75bf475c779d2599776f8a83acc08e5