Updated openstack/openstack Project: openstack/keystone  3550b473a715aec80d2e1c70784b362f27deb0ea
Fix multiple SQL backend usage validation error

This patch fixes an error in the file-based domain-specific
configuration support where the method to spot multiple SQL
drivers would fail to remember previous instances of this. The
fix is simply to OR the previous cumulative memory of SQL drivers
with the state of the one being requested.

Change-Id: I87314a5806a5652695ef17f74f70165b8e3347c2
Closes-Bug: #1410850