Updated openstack/openstack Project: openstack/trove  1d5303e21287a84f72d771a74e56dc452a895464
Fixes the resize APIs for Vertica-guest

Identified that there was a service vertica-agent still running
after Vertica database was stopped, because of which Volume mount
was not complete and hence resize-volume was failing.

start_db_with_conf_changes implemented to start Vertica to complete
the workflow of resize-instance.

Also, updated the raise exception for bad-cluster-config to raise
RuntimeError, and removed not needed raise from prepare exception
handler.

Change-Id: I8ccef3aef7725b0c7d46fd8851d69d3d80cfa995
Closes-Bug: #1433852