Updated openstack/openstack Project: openstack/taskflow  fd58fd8bb6f30e5cad4fcdcf2e068a9e0e1f612d
Allow ls() to list recursively (using breadth-first)

Make it easier to list the contents of the full in-memory
filesystem by allowing the ls() function to take a recursive
keyword argument (which by default is false).

This makes it easier to analyze the contents of the full in
memory filesystem; saving people from creating similar code
themselves.

Change-Id: I573797945255cb81728e7a86b58768b848110ee7