Updated openstack/openstack Project: openstack/neutron  29effc67fc78186206061b06a0e8f9795e8189b0
Store and log correct exception info

Since OVSDB commands execute in a different thread, the exceptions
that are passed to the original thread do not contain traceback
info from the exception. This patch stores the text from the
exception as it is caught so that the calling thread can log it.

Change-Id: If462c3d5dc104b349218dc910aa281220a5af528