Updated openstack/openstack Project: openstack/horizon  c3fc2678e466e6f139015256a66c3b45c8f7b1ca
Refactoring the exceptions lists

The exception list is not properly used. The exception list is a
settings constant that can be modified within the settings file.  Having
the list in openstack_dashboard/exceptions.py confuses the
developers. The list of exceptions are moved in the settings file and
removed all duplications.

Partially Implements: blueprint improve-error-message-details-for-usability
Change-Id: I5e19c3f0b84b17fa28f6e3c113467cc19d9b0a03