Updated openstack/openstack Project: openstack/cinder  f8b6feb5eb9bcf41a68ad516da520b79559faa4c
Check volume status in detach db api

If a volume is detached while uploading volume to image is in progress then
it sets volume status to available and then it is allowed to delete it
immediately while uploading is still running in background.

Set the status to available in the detach db api only when the volume
is not migrating and status is not uploading.

Closes-Bug: 1413887
Change-Id: I6df9c6918825fba7c8d7cd6521752b399dd61610