Updated openstack/openstack Project: openstack/tripleo-image-elements  25845ab43bfde53b4f42e6f354d16dbe8732727c
Check USE_CACHE for fedora and ubuntu images.

check if use cache for fedora and precise images and query for
fedora image name only if USE_CACHE is false so internet access
is not necessary.
also fixed a bug in setting NP_CREDS, openstack-nodepool usename
and password should be used for its tenant.

Change-Id: I2d9f662b7764a95a7de3e62592f3a8bcb24b3a52