Updated openstack/openstack Project: openstack/barbican  0e076050d63836e00f2c6b3d9599351242d19690
Implement validators and tests for stored key certificate orders

Adding validators for the stored key method of the Certificate
Order API.  Validate the container ref is valid and the
container exists and is of the right type.  Some test
infrastructure code will be useful for other work.

Change-Id: I93ad6b630424795be9d3d4d2b599a7100c4568f9
Implements: blueprint certificate-order-api