Updated openstack/openstack Project: openstack/horizon  f8bbba1067fd49b3669541303a3535823c9602c0
Adding transclude to search-bar for batch actions

Currently, there is no good way to align the batch/table actions along with
the search bar. The best way is to treat each element in the row as a
table-cell. To achieve this, we transclude the search bar so that it is possible
to embed additional actions.

Co-Authored-By: Kelly Domico <kelly.domico@hp.com>
Closes-bug: #1436107
Change-Id: I47905601e4ddf687d6554ff67eabaa3e6985f202