Updated openstack/openstack Project: openstack/tempest  6c353f0843824bbdb6ae1f3f135f672fc18b1809
L3 Agent Scheduler testcase cleanups

This patch cleans up the test case a bit: it uses the base class
method to create/delete routers, and ensure that, when selecting
the L3 agent, this is of a specific type, i.e. legacy. In multi
node environments, and especially the ones post-Juno, there may
be more than one L3 agent running and having the ability to filter
based on the type is instrumental to getting this test case to
work while validating multi-node deployments.

Change-Id: I479b8f52e6da0f520bda965580beddbae0c6a296